feat(common): add obfuscation utilities fix(pds): show user did in logs fix(ozone): show user did in logs
