+++ .changeset/young-ligers-chew.md
@@ -1,0 +1,5 @@
+---
+"@atproto/common": minor
+---
+
+Add obfuscation utilities
